+++ CHANGES.md
@@ -13,6 +13,12 @@
 - [CHANGE] spotlight_legacy 対応を削除する
     - ConnectionOptions の spotlight オプションの型を boolean のみに変更する
 
+## 2021.1.6
+- [FIX] timeline ログに re-answer のログが出力されていなかったので修正する
+    - @yuitowest
+- [UPDATE] timeline ログの ontrack ログに詳細情報を追加する
+    - @yuitowest
+
 ## 2021.1.5
 - [FIX] dataChannelSignaling false の場合に Disconnect API 経由で切断すると disconnect callback が発火しない問題を修正する
     - @yuitowest